+++ config/urls.py
@@ -23,6 +23,7 @@ from wies.core.views import AssignmentDetailView, ColleagueDetailView
 from wies.core.views import PlacementTableView
 from wies.core.views import MinistryDetailView
 from wies.core.views import admin_db, login, no_access, logout, auth
+from wies.core.views import UserListView, user_create, user_delete
 
 urlpatterns = [
     path('admin/db/', admin_db, name='admin-db'),
@@ -37,4 +38,7 @@ urlpatterns = [
     path('placements/', PlacementTableView.as_view(), name='placements'),
     path('clients/<str:name>', client),
     path('ministries/<int:pk>/', MinistryDetailView.as_view(), name='ministry-detail'),
+    path('users/', UserListView.as_view(), name='users'),
+    path('users/create/', user_create, name='user-create'),
+    path('users/<int:pk>/delete/', user_delete, name='user-delete'),
 ]